HRMS-Server/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
boto3==1.29.6
botocore==1.32.6
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudinary==1.36.0
colorama==0.4.6
cryptography==41.0.5
cssselect2==0.7.0
Django==4.2.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
freetype-py==2.3.0
gevent==23.9.1
greenlet==3.0.1
grequests==0.7.0
html5lib==1.1
idna==3.4
jmespath==1.0.1
lxml==4.9.3
oscrypto==1.3.0
pdfkit==1.0.0
Pillow==10.0.1
psycopg==3.1.13
psycopg2==2.9.9
pycairo==1.25.1
pycparser==2.21
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyJWT==2.7.0
pypdf==3.17.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.7
requests==2.31.0
rlPyCairo==0.3.0
s3transfer==0.7.0
setuptools==69.0.2
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.2
urllib3==1.26.5
webencodings==0.5.1
xhtml2pdf==0.2.13
zope.event==5.0
zope.interface==6.1